Fix elementFactory

+++ src/RealmCore.Server/Modules/Domain/Location.cs
@@ -11,7 +11,7 @@ public record Location(Vector3 Position, Vector3 Rotation = default, byte? Inter
     public static Location Zero => _location;
 
     public byte GetInteriorOrDefault() => Interior ?? 0;
-    public byte GetDimensionOrDefault() => Dimension ?? 0;
+    public ushort GetDimensionOrDefault() => Dimension ?? 0;
 
     public override string ToString()
     {
